676 - fix hmr
