+++ src/scripts/index.js
@@ -2,35 +2,22 @@
 
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { AppContainer } from 'react-hot-loader'
-import { Provider } from 'react-redux'
 import 'promise-polyfill/src/polyfill'
 
 import Root from 'components/Root'
+import PortalWrapper from 'components/PortalWrapper'
 import { getRoot } from 'utils/AppUtils'
 import initializeTranslator from 'utils/translations/initializeTranslator'
-import TranslationContext from 'utils/translations/TranslationContext'
 import createStore from 'scripts/createStore'
+
 import 'styles/app.scss'
 
 const store = createStore()
 
 const translator = initializeTranslator()
 
-type Props = {
-  children: any
-}
-
-const PortalWrapper = ({ children }: Props) => (
-  <TranslationContext.Provider value={translator}>
-    <Provider store={store}>
-      <AppContainer>{children}</AppContainer>
-    </Provider>
-  </TranslationContext.Provider>
-)
-
 ReactDOM.render(
-  <PortalWrapper>
+  <PortalWrapper store={store} translator={translator}>
     <Root />
   </PortalWrapper>,
   getRoot()
@@ -40,7 +27,7 @@ if (module.hot) {
   module.hot.accept('components/Root', () => {
     const NextRoot = require('components/Root').default
     ReactDOM.render(
-      <PortalWrapper>
+      <PortalWrapper store={store} translator={translator}>
         <NextRoot />
       </PortalWrapper>,
       getRoot()
